requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.4.1
astroid==2.0.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
cffi==1.15.0
chardet==3.0.4
colorama==0.4.3
cryptography==3.4.8
cycler==0.10.0
defusedxml==0.5.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.45.0
django-cors-headers==3.7.0
django-countries==7.2.1
django-heroku==0.3.1
django-location-field==2.1.0
django-phone-field==1.8.1
django-phonenumber-field==5.2.0
django-rest-auth==0.9.5
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
idna==2.7
isort==4.3.4
kiwisolver==1.3.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
phonenumbers==8.12.31
Pillow==8.4.0
psycopg==3.0.1
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
pynput==1.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.1.0
pytz==2021.1
requests==2.21.0
requests-oauthlib==1.0.0
six==1.11.0
soupsieve==2.2
sqlparse==0.4.1
tk==0.1.0
toml==0.10.1
tzlocal==2.1
urllib3==1.24.2
virtualenv==20.4.2
whitenoise==5.2.0
wrapt==1.10.11
